# services/core-go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.UserRole:
    enum:
    - super_admin
    - admin
    - finance
    - direktur
    - owner
    - operator
    - viewer
    type: string
    x-enum-comments:
      RoleAdmin: Administrator Bisnis / Head of Operations
      RoleDirektur: Direktur / Manajemen (Monitoring & Review)
      RoleFinance: Staf Keuangan & Akuntansi (Input Transaksi & Invoice)
      RoleOwner: Pemilik Modal / Investor (Executive Dashboard)
      RoleSuperAdmin: IT Master / Developer (Tersembunyi, Fail-safe All Access)
    x-enum-descriptions:
    - IT Master / Developer (Tersembunyi, Fail-safe All Access)
    - Administrator Bisnis / Head of Operations
    - Staf Keuangan & Akuntansi (Input Transaksi & Invoice)
    - Direktur / Manajemen (Monitoring & Review)
    - Pemilik Modal / Investor (Executive Dashboard)
    - ""
    - ""
    x-enum-varnames:
    - RoleSuperAdmin
    - RoleAdmin
    - RoleFinance
    - RoleDirektur
    - RoleOwner
    - RoleOperator
    - RoleViewer
  domain.UserStatus:
    enum:
    - ACTIVE
    - INACTIVE
    - SUSPENDED
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusInactive
    - StatusSuspended
...
  handler.IdempotencyKeyResponse:
    properties:
      expires_in_hours:
        example: 24
        type: integer
      idempotency_key:
        example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        type: string
      usage_hint:
        example: Include this key in the 'X-Idempotency-Key' request header for POST/PUT/PATCH/DELETE
          mutations
        type: string
    type: object
...
  handler.ResetPasswordRequest:
    properties:
      password:
        type: string
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: boolean
    type: object
  response.PaginationMeta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  services.CreateUserInput:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/domain.UserRole'
      status:
        $ref: '#/definitions/domain.UserStatus'
    type: object
...
  services.UpdateUserInput:
    properties:
      email:
        type: string
      full_name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/domain.UserRole'
      status:
        $ref: '#/definitions/domain.UserStatus'
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    High Performance Backend Service for PT. Adijayantara Logistics Indonesia.
    Architecture & Headers:
    - X-API-Version: Target API versioning (e.g. 'v1', 'v2')
    - X-Idempotency-Key: Unique UUID token for safe mutations (POST/PUT/PATCH/DELETE) to prevent duplicates
    - X-Request-ID: Distributed tracing correlation ID
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Cashflow & Shipment Management API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Retrieves list of users with pagination, search, role, and status
        filtering
      parameters:
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 15
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by name, email, or phone
        in: query
        name: search
        type: string
      - description: Filter by role code
        in: query
        name: role
        type: string
      - description: Filter by status (ACTIVE, INACTIVE, SUSPENDED)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Registers a new staff user with designated role
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: New User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Soft deletes or deactivates a user (enforces Anti Self-Deletion,
        Last Admin Standing, and C-Level Immunity)
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Soft delete / deactivate user
      tags:
      - users
    get:
      description: Retrieves profile of a specific user
      parameters:
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates profile, status, or role of an existing user (enforces
        C-Level & Self-Deletion Golden Rules)
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/{id}/password:
    patch:
      consumes:
      - application/json
      description: Resets the password for a user account
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: New Password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Reset user password
      tags:
      - users
  /utility/idempotency-key:
    get:
      description: Generates a new RFC 4122 UUID v4 to be used as 'X-Idempotency-Key'
        header on mutation endpoints (POST/PUT/PATCH/DELETE).
      parameters:
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.IdempotencyKeyResponse'
              type: object
      summary: Generate fresh Idempotency Key
      tags:
      - utility
...
securityDefinitions:
  ApiVersion:
    description: Target API Version Routing (e.g. v1, v2)
    in: header
    name: X-API-Version
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and the session token.
    in: header
    name: Authorization
    type: apiKey
  IdempotencyKey:
    description: Unique UUID idempotency key for mutation operations (POST/PUT/PATCH/DELETE)
    in: header
    name: X-Idempotency-Key
    type: apiKey
swagger: "2.0"